feat: scroll-triggered node burst on landing CTA section
